Update index.js
Sometimes file system will take a long time to write the file on disk. It is not recommended to abort the rest of the process as the file will finally witten on disk and should be renamed to the final name. Another strategy could be to rewrite getOutputFilePath() function to directly save the file on disk with the final name and avoid the renaming process.